+++ src/main/java/pl/grizzlysoftware/chlorek/core/model/ProductIngredient.java
@@ -1,11 +1,13 @@
 package pl.grizzlysoftware.chlorek.core.model;
 
+import java.math.BigDecimal;
+
 /**
  * @author Bartosz Pawłowski, bpawlowski@grizzlysoftware.pl
  */
 public class ProductIngredient {
     public Long id;
     public Long productId;
-    public Integer quantity;
+    public BigDecimal quantity;
     public String unit;
 }
